feat: add SHORT_ROUTER_URL to the build environment
The domain name of this URL is (almost) guaranteed to be less than 63
chars in length, so it can be used as a CN in certificate requests.

The use of base32 is an attempt to increase the entropy of the 8
character string over simple hex encoding, in order to reduce the
likelihood of collisions. Not sure how necessary it is, really.